refactor: use focused test file for CSS import attribute regression tests
- Moved tests to separate file to avoid interfering with existing import-attributes tests
- Tests specifically target transpiler behavior for issue #15310
- All tests pass and verify the fix works correctly
